Revert "uint16"
This reverts commit ca99a3a3f70ea028d86e275d59d779b59736484f.

+++ src/neurodsp/spiketrains.py
@@ -38,7 +38,7 @@ def spikes_venn3(
                 channels_binsize,
                 [0, int(max_samples)],
                 [0, num_channels],
-            )[0].flatten().astype(np.uint16)
+            )[0].flatten()
             for i in range(3)
         ]
     )
